bs refactor: combine check-types and check-build-system
