+++ build-system/tasks/pr-check.js
@@ -63,10 +63,6 @@ async function prCheck() {
     runCheck('amp ava');
   }
 
-  if (buildTargetsInclude(Targets.BUILD_SYSTEM)) {
-    runCheck('amp check-build-system');
-  }
-
   if (buildTargetsInclude(Targets.BABEL_PLUGIN)) {
     runCheck('amp babel-plugin-tests');
   }
